Buttons and alerts added
The alerts are not meant to be stacked, but they won’t cause any issues
if this happens. The status bar may appear a bit prematurely, that’s
all.

These alerts are with the backgroundColor and textColor arguments set
to nil, which uses the default colors shown. You can send any
background / textColor combo you require.